requirements.txt
asgiref==3.7.2
CacheControl==0.13.1
cachetools==5.3.2
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
cryptography==41.0.7
defusedxml==0.8.0rc2
Django==4.2.2
django-activity-stream==2.0.0
django-autoslug==1.9.9
django-cities-light==3.10.1
django-ckeditor-5==0.2.10
django-crispy-forms==2.1
django-jazzmin==2.6.0
django-phonenumber-field==7.3.0
django-phonenumbers==1.0.1
django-push-notifications==3.0.2
firebase-admin==6.3.0
google-api-core==2.15.0
google-api-python-client==2.114.0
google-auth==2.26.2
google-auth-httplib2==0.2.0
google-cloud-core==2.4.1
google-cloud-firestore==2.14.0
google-cloud-storage==2.14.0
google-crc32c==1.5.0
google-resumable-media==2.7.0
googleapis-common-protos==1.62.0
grpcio==1.60.0
grpcio-status==1.60.0
httplib2==0.22.0
idna==3.6
msgpack==1.0.7
oauthlib==3.2.2
phonenumbers==8.13.27
pillow==10.2.0
progressbar2==4.3.2
proto-plus==1.23.0
protobuf==4.25.2
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycparser==2.21
PyJWT==2.8.0
pyparsing==3.1.1
python-dotenv==1.0.0
python-utils==3.8.1
python3-openid==3.2.0
pytz==2023.3.post1
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
social-auth-app-django==5.4.0
social-auth-core==4.5.1
sqlparse==0.4.4
typing_extensions==4.9.0
Unidecode==1.3.8
uritemplate==4.1.1
urllib3==2.1.0
whitenoise==6.6.0
